requirements.txt
Flask
flask-injector
flask-sqlalchemy
python-dotenv